+++ packages/payloads/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.4.10](https://github.com/standardnotes/snjs/compare/@standardnotes/payloads@1.4.9...@standardnotes/payloads@1.4.10) (2022-03-14)
+
+**Note:** Version bump only for package @standardnotes/payloads
+
+
+
+
+
 ## [1.4.9](https://github.com/standardnotes/snjs/compare/@standardnotes/payloads@1.4.8...@standardnotes/payloads@1.4.9) (2022-03-11)
 
 **Note:** Version bump only for package @standardnotes/payloads